+++ lib/Phpfastcache/Core/Pool/DriverBaseTrait.php
@@ -143,7 +143,7 @@ trait DriverBaseTrait
     {
         $wrap = [
             self::DRIVER_KEY_WRAPPER_INDEX => $item->getKey(), // Stored but not really used, allow you to quickly identify the cache key
-            self::DRIVER_DATA_WRAPPER_INDEX => $item->get(),
+            self::DRIVER_DATA_WRAPPER_INDEX => $item->getRawValue(),
             self::DRIVER_TAGS_WRAPPER_INDEX => $item->getTags(),
             self::DRIVER_EDATE_WRAPPER_INDEX => $item->getExpirationDate(),
         ];