+++ tests/Firestore.test.php
@@ -14,35 +14,32 @@
 
 use Phpfastcache\CacheManager;
 use Phpfastcache\Core\Pool\ExtendedCacheItemPoolInterface;
-use Phpfastcache\Drivers\Dynamodb\Config as DynamodbConfig;
+use Phpfastcache\Drivers\Firestore\Config as FirestoreConfig;
 use Phpfastcache\Exceptions\PhpfastcacheDriverConnectException;
 use Phpfastcache\Tests\Helper\TestHelper;
 
 chdir(__DIR__);
 require_once __DIR__ . '/../vendor/autoload.php';
-$testHelper = new TestHelper('Amazon Dynamodb driver');
+$testHelper = new TestHelper('Google Firestore driver');
 
-$config = new DynamodbConfig();
+/**
+ * This driver awaits some fixes on Google side related to psr/cache version
+ *
+ * @see https://github.com/googleapis/google-auth-library-php/pull/364
+ * @see https://github.com/googleapis/google-auth-library-php/issues/363
+ */
+
+
+
+/*$config = new FirestoreConfig();
 
 try {
     $config->setItemDetailedDate(true);
-    $config->setRegion('eu-west-3');
-    $config->setEndpoint('dynamodb.eu-west-3.amazonaws.com');
-    $config->setTable('phpfastcache');
-    $cacheInstance = CacheManager::getInstance('Dynamodb', $config);
-
-    $cacheInstance->getEventManager()->onDynamodbCreateTable(static function(ExtendedCacheItemPoolInterface $pool, array $params) use ($testHelper){
-        $testHelper->printDebugText(
-            sprintf(
-                'Table created with the following parameters: %s',
-                json_encode($params, JSON_THROW_ON_ERROR)
-            )
-        );
-    });
-
+    $config->setCollection('phpfastcache');
+    $cacheInstance = CacheManager::getInstance('Firestore', $config);
     $testHelper->runCRUDTests($cacheInstance, false);
 } catch (PhpfastcacheDriverConnectException $e) {
     $testHelper->assertSkip('Dynamodb server unavailable: ' . $e->getMessage());
     $testHelper->terminateTest();
-}
+}*/
 $testHelper->terminateTest();